requirements.txt
absl-py==2.1.0
arch==7.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
bleach==6.1.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.6
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
html5lib==1.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kaggle==1.6.17
kagglehub==0.3.3
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mplfinance==0.12.10b0
mpmath==1.3.0
multitasking==0.0.11
namex==0.0.8
networkx==3.4.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
patsy==0.5.6
peewee==3.17.7
pillow==11.0.0
platformdirs==4.3.6
pmdarima==2.0.4
protobuf==5.28.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soupsieve==2.6
statsmodels==0.14.4
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.1
wrapt==1.16.0
yfinance==0.2.45
